gnt-node-shutdown.sh for graceful shutdown with UPS
When 2 paired nodes going down together there is the
big chance that the secondary node will go down first,
and the primary node's instances will go down without
sync to secondary.

This script causes instances shutdown on the node,
kills ganeti daemon and lets node to be up to receive
all coming drbd requests.

This is like a hack, but seems to be the better
practical solution till now.